feat: modular --post commands
These can be used to do arbitrary post-processing on the output from the
endpoint. It can be used to just output the plain text, as is the
default, or put it into a markdown file and try to open it, etc.

The only requirement for a post script is that it is bash-compatible.
The return code from the post script is what clai will return.